Refactor scrolltop to common ui
