+++ apps/frontend/src/app/App.tsx
@@ -2,6 +2,7 @@ import {
   DBLocalStorage,
   SandboxStorage,
 } from '@genshin-optimizer/common/database'
+import { ScrollTop } from '@genshin-optimizer/common/ui'
 import { ArtCharDatabase } from '@genshin-optimizer/gi/db'
 import { DatabaseContext } from '@genshin-optimizer/gi/db-ui'
 import '@genshin-optimizer/gi/i18n' // import to load translations
@@ -13,19 +14,15 @@ import {
   useSnow,
   useTitle,
 } from '@genshin-optimizer/gi/ui'
-import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp'
 import {
   Box,
   Container,
   CssBaseline,
-  Fab,
   Skeleton,
   StyledEngineProvider,
   ThemeProvider,
-  Zoom,
-  useScrollTrigger,
 } from '@mui/material'
-import React, { Suspense, lazy, useCallback, useMemo, useState } from 'react'
+import { Suspense, lazy, useCallback, useMemo, useState } from 'react'
 import { HashRouter, Route, Routes } from 'react-router-dom'
 import './App.scss'
 import ErrorBoundary from './ErrorBoundary'
@@ -46,39 +43,6 @@ const PageCharacters = lazy(
 )
 const PageTeams = lazy(() => import('@genshin-optimizer/gi/page-teams'))
 const PageTeam = lazy(() => import('@genshin-optimizer/gi/page-team'))
-
-function ScrollTop({ children }: { children: React.ReactElement }) {
-  const trigger = useScrollTrigger({
-    target: window,
-    disableHysteresis: true,
-    threshold: 100,
-  })
-
-  const handleClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    const anchor = (
-      (event.target as HTMLDivElement).ownerDocument || document
-    ).querySelector('#back-to-top-anchor')
-
-    if (anchor) {
-      anchor.scrollIntoView({
-        behavior: 'smooth',
-        block: 'center',
-      })
-    }
-  }
-
-  return (
-    <Zoom in={trigger}>
-      <Box
-        onClick={handleClick}
-        role="presentation"
-        sx={{ position: 'fixed', bottom: 85, right: 16 }}
-      >
-        {children}
-      </Box>
-    </Zoom>
-  )
-}
 
 function App() {
   const dbIndex = parseInt(localStorage.getItem('dbIndex') || '1')
@@ -125,15 +89,7 @@ function App() {
               <ErrorBoundary>
                 <HashRouter basename="/">
                   <Content />
-                  <ScrollTop>
-                    <Fab
-                      color="secondary"
-                      size="small"
-                      aria-label="scroll back to top"
-                    >
-                      <KeyboardArrowUpIcon />
-                    </Fab>
-                  </ScrollTop>
+                  <ScrollTop />
                 </HashRouter>
               </ErrorBoundary>
             </DatabaseContext.Provider>
