+++ libs/common/ui/src/components/ScrollTop.tsx
@@ -1,0 +1,37 @@
+import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp'
+import { Box, Fab, Zoom, useScrollTrigger } from '@mui/material'
+
+export function ScrollTop() {
+  const trigger = useScrollTrigger({
+    target: window,
+    disableHysteresis: true,
+    threshold: 100,
+  })
+
+  const handleClick = (event: React.MouseEvent<HTMLDivElement>) => {
+    const anchor = (
+      (event.target as HTMLDivElement).ownerDocument || document
+    ).querySelector('#back-to-top-anchor')
+
+    if (anchor) {
+      anchor.scrollIntoView({
+        behavior: 'smooth',
+        block: 'center',
+      })
+    }
+  }
+
+  return (
+    <Zoom in={trigger}>
+      <Box
+        onClick={handleClick}
+        role="presentation"
+        sx={{ position: 'fixed', bottom: 85, right: 16 }}
+      >
+        <Fab color="secondary" size="small" aria-label="scroll back to top">
+          <KeyboardArrowUpIcon />
+        </Fab>
+      </Box>
+    </Zoom>
+  )
+}
