+++ libs/common/ui/src/components/index.ts
@@ -11,6 +11,7 @@ export * from './ImgIcon'
 export * from './InfoTooltip'
 export * from './ModalWrapper'
 export * from './NextImage'
+export * from './ScrollTop'
 export * from './SolidColoredToggleButton'
 export * from './SolidToggleButtonGroup'
 export * from './SqBadge'
